requirements.txt
anthropic
arxiv
google-api-python-client
google-auth
google-auth-oauthlib
lightrag-hku
mcp
pymupdf
requests
voyageai
